Updated client (#4161)

+++ apps/dh/api-dh/source/DataHub.WebApi/Clients/ElectricityMarket/V1/ElectricityMarketClient.cs
@@ -1227,7 +1227,7 @@ namespace Energinet.DataHub.WebApi.Clients.ElectricityMarket.v1
         public string? Capacity { get; set; } = default!;
 
         [Newtonsoft.Json.JsonProperty("powerLimitKw", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public int? PowerLimitKw { get; set; } = default!;
+        public double? PowerLimitKw { get; set; } = default!;
 
         [Newtonsoft.Json.JsonProperty("meterNumber", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
         public string? MeterNumber { get; set; } = default!;
